+++ OrderBook/OrderBook.hpp
@@ -9,14 +9,14 @@
 class OrderBook {
 private:
 
-    std::map<int64_t, std::list<Order>, std::greater<int64_t>> BUY;
-    std::map<int64_t, std::list<Order>> SELL;
+    std::map<uint64_t, std::list<Order>, std::greater<uint64_t>> BUY;
+    std::map<uint64_t, std::list<Order>> SELL;
 public:
-    std::unordered_map<int, OrderRef> OrderPointersStore; // orderid, iterator
+    std::unordered_map<uint64_t, OrderRef> OrderPointersStore; // orderid, iterator
     void AddToOrderBook(const Order& order);
 
     // std::optional is used to represent an object that whether it contain a value or not, so here it either return a order or return nulloptional
-    std::optional<Order> CancelOrder(int Orderid);
+    std::optional<Order> CancelOrder(uint64_t Orderid);
     void PrintOrderBook();
 
     //Getters
@@ -27,7 +27,7 @@ public:
         return SELL;
     }
 
-    void RemovePointer(int id) {
+    void RemovePointer(uint64_t id) {
         OrderPointersStore.erase(id);
     }
 };
